FEC-9908 KalturaPlayer for OVP Provider. Kava embedded. IMA Support. Youbora support. Offline support.
Updated the ConfigData object to support both OTT and OVP, removed unneeded data.
Fixed in OTT, documentation and  updated the code accordingly to the change done in the ConfigData.
Added to the OfflineManager, when calling the prepareAsset function, after the metadata was loaded, if the media has DRM content it will register it.